Rewrite everything with separate objectid and hash
'objectid' now identifies an 'entry', i.e. a full set of metadata
associated with a file or directory.

'filehash' is the hash of the contents (file or directory) and is not
associated with metadata directly; multiple entries can use the same
content, they all have a separate set of metadata.

Entry->content association is done with the builtin key 'hash' which
every entry should possess.